+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-15 12:25:19
-// Commit: test(test): implement widget integration tests for test
+// Updated at: 2026-05-15 12:38:21
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '1181';
-  final String lastMessage = 'test(test): implement widget integration tests for test';
-  final DateTime timestamp = DateTime.parse('2026-05-15T12:25:19+05:00');
+  final String commitRef = '1182';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-05-15T12:38:21+05:00');
 }